requirements.txt
mysql-connector-python>=8.3
ttkbootstrap>=1.10